+++ R/dnnModel.R
@@ -22,7 +22,7 @@ NULL
 #' All Deep Learning related steps in the \pkg{digitalDLSorteR} package are
 #' performed by using the \pkg{keras} package, an API in R for \pkg{keras} in
 #' Python available on CRAN. We recommend using the installation guide available
-#' at \url{https://keras.rstudio.com/} in order to set a more customized
+#' at \url{https://tensorflow.rstudio.com/} in order to set a more customized
 #' configuration.
 #'
 #' \strong{Simulation of bulk RNA-Seq profiles 'on the fly'}
@@ -64,19 +64,19 @@ NULL
 #'   \code{num.hidden.layers} argument.
 #' @param activation.fun Activation function to use (\code{'relu'} by default).
 #'   See the
-#'   \href{https://keras.rstudio.com/reference/activation_relu.html}{keras
+#'   \href{https://tensorflow.rstudio.com/reference/keras/activation_relu.html}{keras
 #'   documentation} to know available activation functions.
 #' @param dropout.rate Float between 0 and 1 indicating the fraction of the
 #'   input neurons to drop in layer dropouts (0.25 by default). By default,
 #'   \pkg{digitalDLSorteR} implements 1 dropout layer per hidden layer.
 #' @param loss Character indicating loss function selected for model training
 #'   (\code{'kullback_leibler_divergence'} by default). See the
-#'   \href{https://keras.rstudio.com/reference/loss_mean_squared_error.html}{keras
+#'   \href{https://tensorflow.rstudio.com/reference/keras/loss-functions.html}{keras
 #'    documentation} to know available loss functions.
 #' @param metrics Vector of metrics used to assess model performance during
 #'   training and evaluation (\code{c("accuracy", "mean_absolute_error",
 #'   "categorical_accuracy")} by default). See the
-#'   \href{https://keras.rstudio.com/reference/metric_binary_accuracy.html}{keras
+#'   \href{https://tensorflow.rstudio.com/reference/keras/metric_binary_accuracy.html}{keras
 #'    documentation} to know available performance metrics.
 #' @param scaling How to scale data before training. It may be:
 #'   \code{"standarize"} (values are centered around the mean with a unit